// packages/contentful/src/types/StandardPageByIdQuery.ts
export interface StandardPageByIdQuery_standardPage_header {
  sys: StandardPageByIdQuery_standardPage_header_sys;
  style: any | null;
  type: string | null;
  logo: StandardPageByIdQuery_standardPage_header_logo | null;
  contentCollection: StandardPageByIdQuery_standardPage_header_contentCollection | null;
  secondaryContentCollection: StandardPageByIdQuery_standardPage_header_secondaryContentCollection | null;
}
// ...
export interface StandardPageByIdQuery_standardPage_contentCollection {
  items: (StandardPageByIdQuery_standardPage_contentCollection_items | null)[];
}
// ...
export interface StandardPageByIdQuery_standardPage_footer {
  sys: StandardPageByIdQuery_standardPage_footer_sys;
  copy: string | null;
  style: any | null;
  type: string | null;
  logo: StandardPageByIdQuery_standardPage_footer_logo | null;
  contentCollection: StandardPageByIdQuery_standardPage_footer_contentCollection | null;
  secondaryContentCollection: StandardPageByIdQuery_standardPage_footer_secondaryContentCollection | null;
}

export interface StandardPageByIdQuery_standardPage {
  id: string | null;
  header: StandardPageByIdQuery_standardPage_header | null;
  contentCollection: StandardPageByIdQuery_standardPage_contentCollection | null;
  footer: StandardPageByIdQuery_standardPage_footer | null;
}

export interface StandardPageByIdQuery {
  standardPage: StandardPageByIdQuery_standardPage | null;
}